1，完善TempFile，2，新增AliWebApiQueryBuilder

+++ src/ClownFish.net/Base/Files/TempFile.cs
@@ -22,38 +22,76 @@ public sealed class TempFile : IDisposable
         if( body == null )
             throw new ArgumentNullException(nameof(body));
 
-        string tempPath = Path.GetFullPath(Path.GetTempPath());
-        string filePath = Path.Combine(tempPath, Guid.NewGuid().ToString("N") + ".tmp");
-
+        string filePath = Path.Combine(EnvUtils.GetTempPath(), Guid.NewGuid().ToString("N") + ".tmp");
 
         File.WriteAllBytes(filePath, body);
 
         // 确认文件是否已写入磁盘
-        for( int i = 0; i < 3; i++ ) {
+        CheckFileLength(filePath, body.Length);
+
+        return new TempFile { FilePath = filePath };
+    }
+
+    /// <summary>
+    /// 确认文件是否已写入磁盘
+    /// </summary>
+    /// <param name="filePath"></param>
+    /// <param name="length"></param>
+    private static void CheckFileLength(string filePath, long length)
+    {
+        for( int i = 0; i < 6; i++ ) {
 
             FileInfo fileInfo = new FileInfo(filePath);
-            if( fileInfo.Length < body.Length )
+            if( fileInfo.Length < length )
                 System.Threading.Thread.Sleep(300);
             else
                 break;
+        }
+    }
+
+    /// <summary>
+    /// 创建临时文件
+    /// </summary>
+    /// <param name="stream"></param>
+    /// <returns></returns>
+    public static TempFile CreateFile(Stream stream)
+    {
+        if( stream == null )
+            throw new ArgumentNullException(nameof(stream));
+
+        if( stream.CanSeek )
+            stream.Position = 0;
+
+        string filePath = Path.Combine(EnvUtils.GetTempPath(), Guid.NewGuid().ToString("N") + ".tmp");
+
+        using( FileStream fileStream = File.Create(filePath) ) {
+            stream.CopyTo(fileStream);
+        }
+
+        if( stream.CanSeek ) {
+            CheckFileLength(filePath, stream.Length);
         }
 
         return new TempFile { FilePath = filePath };
     }
 
+
     [SuppressMessage("Microsoft.Design", "CA1063:ImplementIDisposableCorrectly")]
     void IDisposable.Dispose()
     {
         // 删除临时文件
+        if( File.Exists(this.FilePath) ) {
 
-        for( int i = 0; i < 10; i++ ) {
-            try {
-                File.Delete(this.FilePath);
-            }
-            catch {
-                // 文件有可能没有及时关闭
-                // 忽略所有异常
-                System.Threading.Thread.Sleep(500);
+            for( int i = 0; i < 10; i++ ) {
+                try {
+                    File.Delete(this.FilePath);
+                    return;
+                }
+                catch {
+                    // 文件有可能没有及时关闭
+                    // 忽略所有异常
+                    System.Threading.Thread.Sleep(300);
+                }
             }
         }
     }
